fold together like minded logic and test
