+++ src/primitives/index.ts
@@ -5,4 +5,3 @@ export * from "./storyboards";
 export * from "./text-chunking";
 export * from "./thumbnails";
 export * from "./transcripts";
-export * from "./vtt-chunking";
